# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trade Analytics Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, all]
        - name: setup
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /journal/entries:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /journal/tags:
    get:
      operationId: listJournalTags
      tags: [journal]
      summary: List all journal tags
      responses:
        "200":
          description: Tags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalTag"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /insights/weekly-review:
    get:
      operationId: getWeeklyReview
      tags: [insights]
      summary: Get weekly performance review
      responses:
        "200":
          description: Weekly review data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReview"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /insights/alerts:
    get:
      operationId: getCoachingAlerts
      tags: [insights]
      summary: Get coaching alerts and recommendations
      responses:
        "200":
          description: Coaching alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachingAlert"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /insights/strategy-drift:
    get:
      operationId: getStrategyDrift
      tags: [insights]
      summary: Get strategy drift analysis
      responses:
        "200":
          description: Strategy drift data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyDrift"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /brokers:
    get:
      operationId: listBrokers
      tags: [brokers]
      summary: List all connected brokers
      responses:
        "200":
          description: Broker connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerConnection"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /brokers/sync-status:
    get:
      operationId: getBrokerSyncStatus
      tags: [brokers]
      summary: Get broker sync status
      responses:
        "200":
          description: Sync status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /portfolio/overview:
    get:
      operationId: getPortfolioOverview
      tags: [portfolio]
      summary: Get portfolio overview across all brokers
      responses:
        "200":
          description: Portfolio overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthUser:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        avatarInitials:
          type: string
      required: [username, displayName, avatarInitials]
...
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        status:
          type: string
          enum: [open, closed]
        setup:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        pnl:
          type: number
        pnlPercent:
          type: number
        rr:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        broker:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [id, symbol, side, status, setup, entryPrice, quantity, pnl, openTime, broker, tags]
    TradeList:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
      required: [trades, total]
    TradeDetail:
      allOf:
        - $ref: "#/components/schemas/Trade"
        - type: object
          properties:
            notes:
              type: string
            screenshots:
              type: array
              items:
                type: string
            stopLoss:
              type: number
            takeProfit:
              type: number
            mae:
              type: number
            mfe:
              type: number
            commission:
              type: number
...
    JournalEntry:
      type: object
      properties:
        id:
          type: string
        tradeId:
          type: string
        symbol:
          type: string
        date:
          type: string
        pnl:
          type: number
        rr:
          type: number
        setup:
          type: string
        emotion:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: integer
      required: [id, tradeId, symbol, date, pnl, rr, setup, notes, tags]
    JournalTag:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        color:
          type: string
      required: [name, count, color]
    WeeklyReview:
      type: object
      properties:
        weekStart:
          type: string
        weekEnd:
          type: string
        totalPnL:
          type: number
        totalTrades:
          type: integer
        winRate:
          type: number
        avgRR:
          type: number
        bestDay:
          type: string
        worstDay:
          type: string
        topSetup:
          type: string
        improvementAreas:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        dailyBreakdown:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              pnl:
                type: number
              trades:
                type: integer
            required: [day, pnl, trades]
      required:
        - weekStart
        - weekEnd
        - totalPnL
        - totalTrades
        - winRate
        - avgRR
        - bestDay
        - worstDay
        - topSetup
        - improvementAreas
        - strengths
        - dailyBreakdown
    CoachingAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [warning, tip, achievement, risk]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
        dismissed:
          type: boolean
      required: [id, type, title, message, severity, createdAt, dismissed]
    StrategyDrift:
      type: object
      properties:
        driftScore:
          type: number
        baselineSetup:
          type: string
        currentTopSetup:
          type: string
        setupDistribution:
          type: array
          items:
            type: object
            properties:
              setup:
                type: string
              percentage:
                type: number
              change:
                type: number
            required: [setup, percentage, change]
        weeklyDrift:
          type: array
          items:
            type: object
            properties:
              week:
                type: string
              score:
                type: number
            required: [week, score]
        recommendation:
          type: string
      required:
        - driftScore
        - baselineSetup
        - currentTopSetup
        - setupDistribution
        - weeklyDrift
        - recommendation
    BrokerConnection:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [connected, disconnected, syncing, error]
        accountBalance:
          type: number
        openPositions:
          type: integer
        lastSync:
          type: string
        logo:
          type: string
        accountId:
          type: string
      required: [id, name, status, accountBalance, openPositions, lastSync, accountId]
    SyncStatus:
      type: object
      properties:
        lastSyncAt:
          type: string
        nextSyncAt:
          type: string
        tradesImported:
          type: integer
        brokers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              status:
                type: string
              tradesImported:
                type: integer
              lastSync:
                type: string
            required: [name, status, tradesImported, lastSync]
      required: [lastSyncAt, nextSyncAt, tradesImported, brokers]
    PortfolioOverview:
      type: object
      properties:
        totalBalance:
          type: number
        totalPnL:
          type: number
        totalPnLPercent:
          type: number
        openPositions:
          type: integer
        allocations:
          type: array
          items:
            type: object
            properties:
              broker:
                type: string
              balance:
                type: number
              percentage:
                type: number
            required: [broker, balance, percentage]
        positions:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              side:
                type: string
              quantity:
                type: number
              entryPrice:
                type: number
              currentPrice:
                type: number
              unrealizedPnL:
                type: number
              broker:
                type: string
            required: [symbol, side, quantity, entryPrice, currentPrice, unrealizedPnL, broker]
      required: [totalBalance, totalPnL, totalPnLPercent, openPositions, allocations, positions]